Add null checks, hopefully fixes issue.

+++ src/api/java/WayofTime/alchemicalWizardry/api/items/ShapedBloodOrbRecipe.java
@@ -1,11 +1,6 @@
 package WayofTime.alchemicalWizardry.api.items;
 
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.Iterator;
-import java.util.Map;
-import java.util.Map.Entry;
-
+import WayofTime.alchemicalWizardry.api.items.interfaces.IBloodOrb;
 import net.minecraft.block.Block;
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.Item;
@@ -14,7 +9,12 @@ import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.item.crafting.ShapedRecipes;
 import net.minecraft.world.World;
 import net.minecraftforge.oredict.OreDictionary;
-import WayofTime.alchemicalWizardry.api.items.interfaces.IBloodOrb;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Iterator;
+import java.util.Map;
+import java.util.Map.Entry;
 
 /**
  * Shaped Blood Orb Recipe Handler by joshie *
@@ -102,8 +102,8 @@ public class ShapedBloodOrbRecipe implements IRecipe
             if (in instanceof IBloodOrb || (in instanceof ItemStack && ((ItemStack) in).getItem() instanceof IBloodOrb))
             { //If the item is an instanceof IBloodOrb then save the level of the orb
                 if (in instanceof ItemStack)
-                    itemMap.put(chr, (Integer) (((IBloodOrb) ((ItemStack) in).getItem()).getOrbLevel()));
-                else itemMap.put(chr, (Integer) (((IBloodOrb) in).getOrbLevel()));
+                    itemMap.put(chr, ((IBloodOrb) ((ItemStack) in).getItem()).getOrbLevel());
+                else itemMap.put(chr, ((IBloodOrb) in).getOrbLevel());
             } else if (in instanceof ItemStack)
             {
                 itemMap.put(chr, ((ItemStack) in).copy());
